Ref #133: Use just "ps" to list processes
This doesn't fix the problem for Android O as I initially thought
because new ps implementation only lists processes associated with a
current tty (most of the time, - running "adb shell ps" lists all
processes except logcats spawned by the logview but
"adb shell <Enter> ps" lists only sh and ps itself). For Android O+ the
"ps -A" has to be used but it is incompatible with previous versions. I
plan to introduce version-specific process dumpers in the next commit.

Anyway "-P" is unnecessary even on previous versions because fg/bg
priority was never of concern. It was used for compatibility with
dumpstate output (it used ps -P prior to M).

This commit adds comprehensive test suite of sample ps outputs from
2.3.3 to P (9.0?) gathered from emulator.