Fix #141: Replace Buffer.UNKNOWN with plain null
Having UNKNOWN value in the enum surprisingly makes code more
error-prone because not all users of the enum take in account that
UNKNOWN is special. They just iterate through values(). Null is
safer and cleaner option in this case. With this CL:
 - attempt to read 'unknown' buffer from device is gone
 - no NPE when setting available menu item if AdbDataSource decides
   that unknown buffer is actually available.

The latter will be fixed during refactorings of the adb-communicating
code.