Add a new scope: LogTableScoped, per log table instance
The ultimate goal is to allow using scopes and DI in .ui.logtable
package. Users of LogTable should create a separate component (like
MainLogTableComponent and IndexLogTableComponent), annotate it with
@LogTableScoped and use the component to retrieve LogTable instance.
This trick enables new architecture possibility that will be used in the
follow-up CLs. Suppose that logtable package exposes public interfaces
Service and Dependency. There is a scoped, so one-per-table,
private implementation of Service in logtable package that is binded to
the interdace in dagger Module. Logtable also requires Dependency to be
provided via DI. With the trick a client can provide implementation of
dependency that depends on Service interface and get a scoped instace of
that Service. There is no obvious way to achieve this otherwise, without
providing scope in logtable and assembling a final Component on the
client side.