Optimize performance of trivial and/or predicates
There is no need to introduce complex predicate chain if the predicate
is trivial (true for and, false for or) or if there is only one
predicate to combine - it can be used as is.

The JMH benchmark shows 30% improvement of processing speed on my
machine though I didn't do thorough analysis.