Add parser for the Android Studio logs
Logs saved from the logcat window of the Android Studio are sometimes
encountered in the wild, so it makes sense to be able to read them. Most
notable distinction is that these logs may contain app name.

Current implementation assumes that app name has no spaces in it, this
might be wrong but I'm not sure if it is possible to handle reliably.