Introduce new DDMLIB facade
AdbManager is the primary entry point. It is used to obtain an instance
of the AdbServer which represents active connection and will be used to
do all ADB-related tasks. The split is to be able to create AdbManager
without touching the DDMLIB machinery.

The AdbConnection is a temporary class that exposed the current active
AndroidDebugBridge. It will go away eventually if I don't figure out
more uses for it.

Issue: #75