Use consistent file permissions in archives everywhere
Without explicit set up the build picks up whatever is set on the files
in the working directory. Turns out it depends on the configuration of
the OS, e.g. on Windows and in the Docker image these are `rw?r-?r-?`
while on my Linux box these are `rw?rw?r-?` (executability depends on
the actual file). This causes hard-to-debug discrepancies so now the
build explicitly discards group-writable attribute.

Issue: #157