Fix Filter Panel display when going to/from empty panel
The JPanel's default FlowLayout adds 5px paddings on top and bottom.
Without any filter buttons it is of no concern as the panel with(out)
buttons has a desired height of 0, so it is stretched out to the size of
the container. When the filter button appears, these paddings are
applied and this caused the FilterPanel to become 10px taller.

The fix is to remove the default paddings from the inner JPanel.
Instead, some paddings were added to the FilterPanel itself to achieve
a more pleasant look.

Additionally, a "revalidate()+repaint()" spell is now used when the
filter is added. Without this, the change in height is not propagated
to the FilterPanel (thought there is no more change in height).

Issue: #218