Add the actual provisioning of the device
We need some properties from a device to set up UI and stuff. However,
accessing the properties is inherently async, so this makes UI layer a
bit hard to deal with, especially considering that the device may go
away every moment. To work around this, necessary properties are
retrieved (in async way) and cached before the device is fully exposed
to the UI layer.