Simplify BufferedListener and cover it with perf tests
Making the listener independent of Swing was tough as it was relying on
Swing timers. However, it turns out the timers aren't necessary at all,
and the implementation can be simplified a lot. Performance should be
more or less on the same level. A performance test is added to verify
that.

The new implementation is always buffering but also tries to flush the
buffer as soon as possible. There is no more immediate mode, but the
delay should be even smaller as flush happens as soon as the EDT is
ready to process the posted event. Overflowing is prevented by the fact
that at most one flush job can be scheduled at a single instant. This
allows for natural pauses for the EDT to process other stuff like user
input.