Prefer richer format in FormatSniffer
The TAG and BRIEF formats are very similar and the former can parse the
latter to some extent, confusing PID with parts of tag. However, BRIEF
is richer, so if something can be parsed as BRIEF, it should be. With
this commit, a richer (==with more available fields) format is preferred
when many formats match.