Make waitForDevice mode survive adb failures and restarts
Prior to this commit, AdbDeviceList could only be obtained after ADB
initialization. However, the GlobalAdbDeviceList is already
independent of the ADB connection, it can be initialized without one and
get the connection later. Allowing to obtain the list without wait
greatly simplified the AdbOpener's wait-for-device flow, and made it
resilient to the adb failure. This is an important step to handle ADB
restarts properly.

As part of the implementation, AdbDataSourceFactory doesn't require ADB
to be obtained as well.

Issue: #197