Make DispatchingDeviceList sequence-confined
After making DispatchingDeviceList serialized, there is no need of a
separate AdbDeviceListImpl. We no longer can receive device update in
background, but we've never needed that anyway. A not so nice
side effect is that multiple Device instances may now exist for the
same IDevice - if multiple DispatchingDeviceList are in play, but this
looks bearable.

Issue: #324