Do not use tricky position calculation in FilterModelImpl
This doesn't work because we don't update other filters when removing.
Fixing this makes removal linear, though it is already because the array
list has to be compacted.

It is unlikely that we're going to have tens of thousands of filters, so
having a linear complexity of any filter operation is fine. The only
exception is the initialization, where adding the filters one by one may
exhibit quadratic behavior. To work around that the loading code is now
rewritten.

Issue: #374