Do not use edit mode to toggle checkboxes in the filter tree
Edit mode has assumptions about the selection mode which are hard to
overcome, and, potentially, may cause other issues. This CL changes
toggle handling to a solution that just tracks clicks on checkboxes.

There is an extra layout to determine where checkbox is, but it should
be fast and only happen for the filter being toggled - not that editor
solution is free from this as well.

Another crucial piece is to use mutable tree nodes in the tree model.
The selection handling keeps track of some TreePath instances and
doesn't update them when the underlying model changes. Immutable filter
instances were part of these TreePath, so enabling/disabling the filter
(which replaces the immutable instance in the model) rendered these
paths invalid, breaking selection tracking.

Issue: fixes #384, fixes #385, fixes #386