Replace FilterCollection with brute force reassembling approach
In most cases it is easier to recompute the data each time the filters
change rather than try to keep the implementations in sync with precise
modifications. The only places that somewhat needs accurate information
are the filter display components.

We may reconsider this if this ever becomes a performance bottleneck.