Rework buffer filtering to be done through model
Prior to this, the buffer filter was working on the side, in the table
model filter. This caused issues with the index filter windows because
they've lost the connection somehow. Now everything is done through the
FilterModel.

The buffer filter is always the first filter in the primary filter
model, but this is only by construction. A better approach, perhaps, is
to add a decorating model instead.

Issue: fixes #377