Consider wide selection when opening the filter popup
The actual node in JTree is small, it is only the FlatUI that draws an
edge-to-edge selection. Normal functions that look up the nodes by point
only consider these small areas, which leads to strange behavior when
a user click on the selection but is presented with the popup menu for
no-filter-selected. It has become more prominent with named filters.

This commit adjusts path conversions used for popup menu display to
account for the wide selection.

Issue: #126