Allow installing the same version
We override the ProductCode produced by jpackage, which is a hash of app
name, vendor, and version, with an autogenerated one. Otherwise, the MSI
installer doesn't allow to overwrite the installation with the package
with the same version. Due to technical limitations, all our nightlies
are going to have the same version.

The cost of that is that the installer is no longer deterministic.

Issue: #436