Download platform-tools from Android SDK repository
We download the SDK's XML manifest file, get the license out of it and
present it to the user. Hopefully, this is enough of a fair use - we
don't redistribute the SDK ourselves and give the user necessary info.

So far, the implementation handles happy path but lacks proper error
handling, like checking if the target directory is empty.

I've added ZIP bomb handling just out of curiosity, it isn't really a
crucial bit.

Issue: #239