Discover another adb location if last discovered is no longer valid
Previously, we commited the discovered location and disabled the auto
discovery for next starts. Now we persist the fact that the stored
location was discovered automatically.

We still persist the discovered location to save time searching for it
and to ensure that we use more or less the same ADB between runs.

One open issue is that the stale auto-discovered path leaks into the
preferences when the user opens the ADB configuration dialog.

Issue: #238