Rewrite legacy configuration handling and streamline import
This only keeps the bare minimum of necessary configuration parsing,
without the fancy utils.properties library. The latter is now finally
gone.

As nothing but import reads the legacy configuration, the loading now
happens later, and is controlled by the importer.

The final touch is making the import one-time by setting a special
preference.

Issue: #510